refactor: payment 설정 패키지 분리
