+++ src/main/java/roomescape/config/payment/PaymentProperty.java
@@ -1,4 +1,4 @@
-package roomescape.config;
+package roomescape.config.payment;
 
 public record PaymentProperty(String vendor, String baseUrl, String secretKey, long connectionTimeoutSeconds, long readTimeoutSeconds) {
 }
